Support of AutoCad Architecture implemented.

+++ Source/FDS2ACAD/Fds2AcadSetupActions/BLL/RegistryHelper.cs
@@ -1,8 +1,8 @@
 namespace Fds2AcadSetupActions.BLL
 {
+    using System;
     using Microsoft.Win32;
     using Properties;
-    using System;
 
     public class RegistryHelper
     {
@@ -11,9 +11,13 @@ namespace Fds2AcadSetupActions.BLL
             return null != Registry.LocalMachine.OpenSubKey(Constants.AutoCadRegistryKey);
         }
 
-        public static void CreateFdsBranch()
+        public static void CreateFdsBranch(string acadRegistryKeyName)
         {
-            var acadApplicationsKey = Registry.LocalMachine.OpenSubKey(Constants.AutoCadApplicationsRegistryKey, true);
+            var acadApplicationsKey = Registry.LocalMachine.OpenSubKey(acadRegistryKeyName, true);
+
+            if (acadApplicationsKey == null)
+                throw new ArgumentException("Check if this acad product version is installed and this registry hive exists", "acadRegistryKeyName");
+
             var fdsKey = acadApplicationsKey.CreateSubKey(Constants.FdsPluginRegistryKey, RegistryKeyPermissionCheck.ReadWriteSubTree);
             fdsKey.SetValue(Constants.DescriptionRegValue, Resources.FdsPluginDescription, RegistryValueKind.String);
             var pathToPluginAssembly = string.Format(Constants.PluginFileSystemLocationPattern,
@@ -29,10 +33,38 @@ namespace Fds2AcadSetupActions.BLL
             fdsCommandsKey.SetValue(Constants.BuildMenuCommandName, Constants.BuildMenuCommandName, RegistryValueKind.String);
         }
 
-        public static void RemoveFdsBranch()
+        public static void RemoveFdsBranch(string acadRegistryKeyName)
         {
-            var fdsPluginRegistryKey = String.Concat(Constants.AutoCadApplicationsRegistryKey, Constants.FdsPluginRegistryKey);
+            var fdsPluginRegistryKey = String.Concat(acadRegistryKeyName, Constants.FdsPluginRegistryKey);
+
+            if (Registry.LocalMachine.OpenSubKey(fdsPluginRegistryKey, true) == null)
+                return;
+
             Registry.LocalMachine.DeleteSubKeyTree(fdsPluginRegistryKey);
         }
     }
 }
+
+// Todo :
+// On installing plugin
+//
+// 1 Get all installed versions form SOFTWARE\Autodesk\AutoCAD key
+//     (R1*.*\ACAD-*00*:***)
+// 2 Get supported versions from predefined dictionary
+//                 var versions = new Dictionary<string, string>
+//                                {
+//                                    {"ACAD-7001:409", "AutoCad"},
+//                                    {"ACAD-7004:419", "AutoCad Architecture"}
+//                                };
+// 3 Populate dialog from this dictionary and allow user to select in wich product version 
+//   should the plugin being installed.
+// 
+// 4 Create additional keys only for selected products.
+// 
+// On plugin unistall 
+//
+// 1 Scan registry and find for what products plugin was installed
+// 2 Show user dialog, from wich he can select products, from wich he wants to uninstall plugin
+// 3 a. If user chose all products - remove additional registry keys from all products hives and unistall binaries
+//   b. User chose to unistall plugin only for some versions - remove additional registry keys from 
+//      chosen products ONLY. DO NOT REMOVE BINARIES (prevent standart unistall).
